Moving Hooks to a different class from Starting Step / Adding code to handle when we are not able to take a screenshot
